requirements.txt
google-genai>=1.0.0
feedparser>=6.0.11
httpx>=0.27.0
pyyaml>=6.0.0
rapidfuzz>=3.6.0
python-dotenv>=1.0.0
tenacity>=8.2.0
functions-framework>=3.0.0
